Fixed the usage of imread flags in image_utils. Also fixed version file so that GH actions don't fail. Also removed white space from GH workflow yamls.
