+++ .github/workflows/docker.gpu.udp.yml
@@ -38,7 +38,7 @@ jobs:
         uses: docker/build-push-action@v4
         with:
           push: true
-          tags: roboflow/roboflow-inference-server-udp-gpu:latest,roboflow/roboflow-inference-server-udp-gpu:${{ env.VERSION}}
+          tags: roboflow/roboflow-inference-server-udp-gpu:latest,roboflow/roboflow-inference-server-udp-gpu:${{env.VERSION}}
           cache-from: type=registry,ref=roboflow/roboflow-inference-server-udp-gpu:cache
           cache-to: type=registry,ref=roboflow/roboflow-inference-server-udp-gpu:cache,mode=max
           platforms: linux/amd64
