+++ .github/workflows/docker.jetson.4.5.0.yml
@@ -38,7 +38,7 @@ jobs:
         uses: docker/build-push-action@v4
         with:
           push: true
-          tags: roboflow/roboflow-inference-server-jetson-4.5.0:latest,roboflow/roboflow-inference-server-jetson-4.5.0:${{ env.VERSION}}
+          tags: roboflow/roboflow-inference-server-jetson-4.5.0:latest,roboflow/roboflow-inference-server-jetson-4.5.0:${{env.VERSION}}
           cache-from: type=registry,ref=roboflow/roboflow-inference-server-jetson-4.5.0:cache
           cache-to: type=registry,ref=roboflow/roboflow-inference-server-jetson-4.5.0:cache,mode=max
           platforms: linux/arm64
